[refactor] 대기열 만료 시간 변경 / [style] 패키지 구조 변경 - service
